services/ai-scraper/requirements.txt
crawl4ai>=0.7.4
pydantic>=2.0
python-dotenv
fastapi
uvicorn[standard]

